# pkg/server/http/middleware/iam/files/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Identity and Access Management API",
        "title": "IAM API",
        "contact": {}
    },
    "basePath": "/",
    "paths": {
...
        "/v1/ldap/groups": {
            "get": {
                "tags": [
                    "ldap"
                ],
                "summary": "Get LDAP groups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Response-array_ldap_LdapGroup"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/v1/ldap/sync": {
            "post": {
                "tags": [
                    "ldap"
                ],
                "summary": "Sync LDAP groups",
                "parameters": [
                    {
                        "description": "force",
                        "name": "Body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/iam.SyncRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ResponseError"
                        }
                    }
                }
            }
        },
        "/v1/ldap/sync/{uid}": {
            "post": {
                "tags": [
                    "ldap"
                ],
                "summary": "Sync LDAP groups",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "force",
                        "name": "Body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/iam.SyncRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ResponseError"
                        }
                    }
                }
            }
        },
        "/v1/ldap/users/{uid}": {
            "get": {
                "tags": [
                    "ldap"
                ],
                "summary": "Get LDAP user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Response-ldap_LdapUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.Message": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "data.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total_item_count": {
                    "type": "integer"
                }
            }
        },
...
        "data.Response-array_ldap_LdapGroup": {
            "type": "object",
            "properties": {
                "message": {
                    "$ref": "#/definitions/data.Message"
                },
                "meta": {
                    "$ref": "#/definitions/data.Meta"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ldap.LdapGroup"
                    }
                }
            }
        },
...
        "data.Response-ldap_LdapUser": {
            "type": "object",
            "properties": {
                "message": {
                    "$ref": "#/definitions/data.Message"
                },
                "meta": {
                    "$ref": "#/definitions/data.Meta"
                },
                "payload": {
                    "$ref": "#/definitions/ldap.LdapUser"
                }
            }
        },
...
        "data.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "$ref": "#/definitions/data.Message"
                }
            }
        },
        "data.ResponseMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "$ref": "#/definitions/data.Message"
                }
            }
        },
...
        "iam.SyncRequest": {
            "type": "object",
            "properties": {
                "force": {
                    "type": "boolean"
                }
            }
        },
...
        "ldap.LdapGroup": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "ldap.LdapUser": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        }
    }
}